scroll-anchoring/basic.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  body { height: 1000px }
  div { height: 100px }
</style>
<div id="block1">abc</div>
<div id="block2">def</div>
<script>
test(() => {
  document.scrollingElement.scrollTop = 150;
  document.querySelector("#block1").style.height = "200px";
  assert_equals(document.scrollingElement.scrollTop, 250);
});
</script>
